Moved ConnectNode component to its own file within the Sidebar folder.
